move all logic related to link shortening to BytebinStorage
